Move string-management classes to memory.hpp
