+++ SettingsWatchdog/SettingsWatchdog.vcxproj
@@ -115,6 +115,7 @@
     <ClInclude Include="errors.hpp" />
     <ClInclude Include="logging.hpp" />
     <ClInclude Include="handles.hpp" />
+    <ClInclude Include="memory.hpp" />
     <ClInclude Include="registry.hpp" />
     <ClInclude Include="string-maps.hpp" />
     <ClInclude Include="warnings.hpp" />
